Replace old extension with the new extension
