+++ src/main/groovy/net/neoforged/gradleutils/GradleUtilsExtension.groovy
@@ -22,146 +22,76 @@ package net.neoforged.gradleutils
 
 import groovy.transform.CompileStatic
 import groovy.transform.PackageScope
+import net.neoforged.gradleutils.specs.VersionSpec
 import org.gradle.api.Action
 import org.gradle.api.Project
 import org.gradle.api.artifacts.repositories.MavenArtifactRepository
-import org.gradle.api.file.Directory
 import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.file.ProjectLayout
 import org.gradle.api.provider.Provider
+import org.gradle.api.provider.ProviderFactory
+import org.gradle.api.tasks.Nested
 
 import javax.inject.Inject
 
 @CompileStatic
 abstract class GradleUtilsExtension {
     private final Project project
-    final DirectoryProperty gitRoot
+    private final Provider<String> calculatedVersion
     @PackageScope
     final Provider<GitInfoValueSource.GitInfo> rawInfo
     private final Provider<Map<String, String>> gitInfo
 
     @Inject
-    GradleUtilsExtension(Project project) {
+    GradleUtilsExtension(Project project, ProjectLayout layout, ProviderFactory providers) {
         this.project = project
-        this.gitRoot = project.objects.directoryProperty().convention(project.layout.projectDirectory)
+        gitRoot.convention(layout.projectDirectory)
+
+        this.calculatedVersion = providers.of(VersionCalculatorValueSource) {
+            it.parameters {
+                it.workingDirectory.set(gitRoot)
+                it.versionConfiguration.set(getVersionConfig())
+            }
+        }
+
         this.rawInfo = project.providers.of(GitInfoValueSource) {
             it.parameters {
                 it.workingDirectory.set(this.gitRoot)
             }
         }
         this.gitInfo = project.objects.mapProperty(String, String)
-                .convention(gitRoot.map({ Directory dir -> GradleUtils.gitInfo(dir.asFile) }))
+                .convention(rawInfo.map { it.gitInfo })
     }
 
-    private Map<String, String> getFilteredInfo(boolean prefix = false, String filter) {
-        if (prefix)
-            filter += '**'
-        final valueSource = project.providers.of(GitInfoValueSource) {
-            it.parameters {
-                it.workingDirectory.set(this.gitRoot)
-                it.tagFilters.add(filter)
+    abstract DirectoryProperty getGitRoot()
+
+    @Nested
+    abstract VersionSpec getVersionConfig()
+
+    Object getVersion() {
+        // This allows lazily calculating the version, only when its needed (someone `toString`s this object)
+        return new Object() {
+            @Override
+            String toString() {
+                return calculatedVersion.get()
             }
         }
-        return valueSource.get().gitInfo
     }
 
-    /**
-     * Returns a version in the form {@code $tag.$offset}, e.g. 1.0.5
-     *
-     * @return a version in the form {@code $tag.$offset}, e.g. 1.0.5
-     */
-    String getTagOffsetVersion() {
-        return GradleUtils.getTagOffsetVersion(getGitInfo())
-    }
-
-    /**
-     * Returns a version in the form {@code $tag.$offset}, e.g. 1.0.5.
-     * The provided filter is used to filter the retrieved tag.
-     *
-     * @param prefix If true, will treat the filter as a prefix.
-     * Defaults to false, which means to treat the filter as a glob pattern.
-     * @param filter A non-null string filter used when retrieving the tag
-     * @return a version in the form {@code $tag.$offset}, e.g. 1.0.5
-     */
-    String getFilteredTagOffsetVersion(boolean prefix = false, String filter) {
-        return GradleUtils.getTagOffsetVersion(getFilteredInfo(prefix, filter))
-    }
-
-    /**
-     * Returns a version in the form {@code $tag.$offset}, optionally with the branch
-     * appended if it is not in the defined list of allowed branches
-     *
-     * @param allowedBranches A list of allowed branches; the current branch is appended if not in this list
-     * @return a version in the form {@code $tag.$offset} or {@code $tag.$offset-$branch}
-     */
-    String getTagOffsetBranchVersion(String... allowedBranches) {
-        return GradleUtils.getTagOffsetBranchVersion(getGitInfo(), allowedBranches)
-    }
-
-    /**
-     * Returns a version in the form {@code $tag.$offset}, optionally with the branch
-     * appended if it is not in the defined list of allowed branches.
-     * The provided filter is used to filter the retrieved tag.
-     *
-     * @param prefix If true, will treat the filter as a prefix.
-     * Defaults to false, which means to treat the filter as a glob pattern.
-     * @param filter A non-null string filter used when retrieving the tag
-     * @param allowedBranches A list of allowed branches; the current branch is appended if not in this list
-     * @return a version in the form {@code $tag.$offset} or {@code $tag.$offset-$branch}
-     */
-    String getFilteredTagOffsetBranchVersion(boolean prefix = false, String filter, String... allowedBranches) {
-        return GradleUtils.getTagOffsetBranchVersion(getFilteredInfo(prefix, filter), allowedBranches)
-    }
-
-    /**
-     * Returns a version in the form {@code $mcVersion-$tag.$offset}, optionally with
-     * the branch appended if it is not in the defined list of allowed branches
-     *
-     * @param mcVersion The current minecraft version
-     * @param allowedBranches A list of allowed branches; the current branch is appended if not in this list
-     * @return a version in the form {@code $mcVersion-$tag.$offset} or {@code $mcVersion-$tag.$offset-$branch}
-     */
-    String getMCTagOffsetBranchVersion(String mcVersion, String... allowedBranches) {
-        return GradleUtils.getMCTagOffsetBranchVersion(getGitInfo(), mcVersion, allowedBranches)
-    }
-
-    /**
-     * Returns a version in the form {@code $mcVersion-$tag.$offset}, optionally with
-     * the branch appended if it is not in the defined list of allowed branches.
-     * The provided filter is used to filter the retrieved tag.
-     *
-     * @param prefix If true, will treat the filter as a prefix.
-     * Defaults to false, which means to treat the filter as a glob pattern.
-     * @param filter A non-null string filter used when retrieving the tag
-     * @param mcVersion The current minecraft version
-     * @param allowedBranches A list of allowed branches; the current branch is appended if not in this list
-     * @return a version in the form {@code $mcVersion-$tag.$offset} or {@code $mcVersion-$tag.$offset-$branch}
-     */
-    String getFilteredMCTagOffsetBranchVersion(boolean prefix = false, String filter, String mcVersion, String... allowedBranches) {
-        return GradleUtils.getMCTagOffsetBranchVersion(getFilteredInfo(prefix, filter), mcVersion, allowedBranches)
+    void version(Action<? extends VersionSpec> configureAction) {
+        configureAction.execute(versionConfig)
     }
 
     Map<String, String> getGitInfo() {
         return gitInfo.get()
     }
 
-    /**
-     * Get a closure to be passed into {@link org.gradle.api.artifacts.dsl.RepositoryHandler#maven(groovy.lang.Closure)}
-     * in a publishing block.
-     *
-     * @param defaultFolder The default folder if the required maven information is not currently set
-     * @return a closure
-     */
-    Action<? extends MavenArtifactRepository> getPublishingForgeMaven(File defaultFolder = project.rootProject.file('repo')) {
+    Action<? extends MavenArtifactRepository> publishingMaven(File defaultFolder = project.rootProject.file('repo')) {
         return GradleUtils.getPublishingForgeMaven(project, defaultFolder)
     }
 
-    /**
-     * Get a closure for the Forge maven to be passed into {@link org.gradle.api.artifacts.dsl.RepositoryHandler#maven(groovy.lang.Closure)}
-     * in a repositories block.
-     *
-     * @return a closure
-     */
-    Action<? extends MavenArtifactRepository> getForgeMaven() {
+    @SuppressWarnings('GrMethodMayBeStatic')
+    Action<? extends MavenArtifactRepository> maven() {
         return GradleUtils.getForgeMaven()
     }
 }
